removed superfluous gui mock element
in practice, users have either imaged a single channel, or all channels that were designated for acquisition at once, i.e. the selection for the snap selection feature never deviated from options now available otherwise. in the same vein, the checkbox to use the autofocus for snap selection lead to confusion and was never used, i.e. when autofocus was avilable, autofocus was also used.