fixed an issue where inactive imaging sites where not skipped, fixed an issue where the autofocus system did not handle the offset sign properly
the code assumed any site was selected, now it checks for the flag properly. to avoid z moves below a precision threshold, the move range was checked, but did not take into account that the offset can be positive or negative where a large positive offset would be considered worthy to move, now both signs are taken into account.